DPV-728: nrf: uart: don't drive TX pin high on init
Driving the TX pin high during pin-initialization for the UART seems has
a nasty side-effect of backpowering the reciever side. This seems to be
sometimes causing incorrect behavior at boot resulting in a dead device
until it gets reset by an operator.
While we mostly care for the initialization step, let's be consistent
and not drive the TX pin high when configuring the GPIOs to be used.

This is OK for our case, since the reciever side includes pull-up
resistors on its RX line (the NRF TX), hence preventing the pin from
floating.

These changes only apply when compiled with NRF_UART_NO_TX_IDLE_DRIVE
defined.